refactor(harness): split CI classification helpers
Lane: L2
Work-Run: f91b7fde-701a-4560-b1c2-b22236046da7
Work-Receipt: g0-r0
