+++ scripts/harness/check-plan.mjs
@@ -1,4 +1,5 @@
 import { classifyScopeChanges, mapFilesToScopes, resolveRequestedScopes } from './shared.mjs';
+import { classifyRepositoryChecks } from './repository-check-classification.mjs';
 
 /**
  * The planned checks that cannot run without the monorepo's build output. CI's `build` job gates
@@ -221,45 +222,6 @@ function addCheck(checks, check) {
   if (!checks.includes(check)) {
     checks.push(check);
   }
-}
-
-function classifyRepositoryChecks(unmappedFiles, rootManifestChange) {
-  const checks = [];
-
-  for (const file of unmappedFiles) {
-    if (file.startsWith('.agents/tasks/')) {
-      addCheck(checks, 'task-plan-scan');
-    } else if (
-      file === 'AGENTS.md' ||
-      file.startsWith('.agents/rules/') ||
-      file.startsWith('.agents/skills/')
-    ) {
-      addCheck(checks, 'harness-consistency');
-      addCheck(checks, 'task-plan-scan');
-    } else if (file.startsWith('scripts/harness/') || file.startsWith('.claude/hooks/')) {
-      addCheck(checks, 'harness-tests');
-      addCheck(checks, 'harness-consistency');
-    } else if (file.startsWith('.github/workflows/') || file.startsWith('.husky/')) {
-      addCheck(checks, 'harness-tests');
-      addCheck(checks, 'harness-consistency');
-    } else if (
-      file === 'package.json' ||
-      file === 'pnpm-workspace.yaml' ||
-      file === 'pnpm-lock.yaml'
-    ) {
-      if (
-        file === 'package.json' &&
-        rootManifestChange?.changedScriptKeys?.some((key) => key.startsWith('harness:'))
-      ) {
-        addCheck(checks, 'harness-tests');
-      }
-      addCheck(checks, 'harness-consistency');
-    } else {
-      addCheck(checks, 'repository-review');
-    }
-  }
-
-  return checks;
 }
 
 export function createVerificationPlan({
