+++ scripts/harness/shared.mjs
@@ -3,6 +3,12 @@ import { appendFileSync, readdirSync, promises as fs } from 'node:fs';
 import path from 'node:path';
 
 import { createBoundedGitRefExists } from './git-base-ref-resolution.mjs';
+import {
+  changedManifestKeys,
+  classifyRootManifestChange,
+} from './manifest-change-classification.mjs';
+
+export { classifyRootManifestChange };
 
 export const WORKSPACE_ROOT = process.cwd();
 const PNPM_WORKSPACE_PATH = path.join(WORKSPACE_ROOT, 'pnpm-workspace.yaml');
@@ -526,28 +532,6 @@ const PACKAGE_PUBLISH_METADATA_FIELDS = [
   'publishConfig',
 ];
 
-function stableJson(value) {
-  if (Array.isArray(value)) {
-    return `[${value.map((item) => stableJson(item)).join(',')}]`;
-  }
-  if (value && typeof value === 'object') {
-    return `{${Object.keys(value)
-      .sort()
-      .map((key) => `${JSON.stringify(key)}:${stableJson(value[key])}`)
-      .join(',')}}`;
-  }
-  return JSON.stringify(value);
-}
-
-function valuesEqual(left, right) {
-  return stableJson(left) === stableJson(right);
-}
-
-function changedManifestKeys(before, after) {
-  const keys = new Set([...Object.keys(before ?? {}), ...Object.keys(after ?? {})]);
-  return Array.from(keys).filter((key) => !valuesEqual(before?.[key], after?.[key]));
-}
-
 export function classifyPackageManifestChange({ before, after }) {
   const changedKeys = changedManifestKeys(before, after);
   const hasVersionOnlyChanges = changedKeys.length === 1 && changedKeys[0] === 'version';
@@ -601,29 +585,6 @@ export function classifyPackageManifestChange({ before, after }) {
     hasPublishMetadataChanges,
     hasUnknownManifestChanges,
     needsSourceHeavyChecks,
-  };
-}
-
-const DEVELOPER_QUALITY_SCRIPT_NAMES = new Set(['lint:fix', 'lint:fix:staged']);
-
-function isDeveloperQualityScript(name) {
-  return DEVELOPER_QUALITY_SCRIPT_NAMES.has(name) || name.startsWith('harness:');
-}
-
-export function classifyRootManifestChange({ before, after }) {
-  const changedKeys = changedManifestKeys(before, after);
-  const changedScriptKeys =
-    changedKeys.length === 1 && changedKeys[0] === 'scripts'
-      ? changedManifestKeys(before?.scripts ?? {}, after?.scripts ?? {})
-      : [];
-  const developerQualityOnly =
-    changedScriptKeys.length > 0 && changedScriptKeys.every((key) => isDeveloperQualityScript(key));
-
-  return {
-    kind: developerQualityOnly ? 'developer-quality-only' : 'workspace-wide',
-    changedKeys,
-    changedScriptKeys,
-    workspaceWide: !developerQualityOnly,
   };
 }
